Tape rotation to address match churn
The previous commit added tests that churns regardless of the implementation of match, replace, etc., and suggest that it isn't possible to have a strict "all of T1, all of T2, etc..." generation order without the possibility of effectively-endless-churning.

This commit turns "tape rotation" back on: instead of finishing T1, then moving onto T2, etc., we move onto the next tape at every step.

Note that tape rotation didn't always work correctly and efficiently in some old versions, but all of those are resolved in recent commits.  (This is just the commit where all those changes happen to have come together for it to work.)  For one, it exacerbated the MatchExpr tape ordering problems, but that's fixed now.  Also I had been putting the rotation at sub-optimal times.  (Rotating BEFORE grabbing the first element as tapeToTry means that you always start on the second tape in the tapePriority, that's probably not quite right.  But meanwhile, rotating within the tightest for loop rotated it too often; you effectively obliterate your semi-carefully chosen priority that way.  The better time is after choosing tapeToTry, and after delta, but before deriv.

This doesn't completely resolve the problems in this branch, though; there's still the issue of replace rules acting as if they're optional.