Generation options extended to Exprs
When trying out different variations of the derivative algorithm, it was getting too complicated to keep track of all the different little things that had to be changed to (for example) switch from LTR to RTL parsing.  Previously we had a GenOptions object, but since it only happened to include options that mattered to pre-deriv operations, Exprs didn't ever use this.  Now it's an argument to deriv.

Also, sampling (via Expr.generateRandom()) was still using the BitSet-based deriv algorithm; this changes it to use the same concreteDeriv() as generation.