New repetition expression with max/min reps
Previously, we had had two different ways of treating repetition:

* For X{0, Infinity} repetition, it was a StarExpr.  (Or currently, RTLStarExpr.)

* For X{m,n} where these are both finite, it was built from concatenations, alternations, and epsilons.  Like X{0,2} would be something like eps|(X+(eps|X)).

For anything else, like {n,Infinity}, it was a combination of these.

When n was big enough, however, we got expressions large enough to blow the stack; the eps|(X+(eps|X+... got too big to recurse through.

So in this commit, I re-did repetition with finite indices in the style of StarExpr.  Now, it's effectively the old StarExpr, except it has minReps and maxReps members; upon "peeling off" an X in deriv(), it decrements both these numbers.  You might expect delta() to check the indices, but it actually is easier to handle all these special cases in constructRepeat().

With this commit, we now pass all tests again.