Replacing binary union exprs with array-based
UnionExprs are used for more than just expressing the grammar, they're also the result of disjoining deriv results.  But since D_c() can potentially have thousands of results even for the same c, we can have UnionExprs of thousands of items.  When UnionExprs are binary, recursing through this can blow the stack.

We had avoided going to an array-based implementation for some time because this stack overflow was actually helpful -- it showed that something else was going wrong.  But now grammar writers are going beyond this threshold with their actual grammars, so it's time to bite the bullet and do it.