Avoiding repeat vocab collection
The longest step of compilation was vocab collection.  Vocab collection requires visiting the referents of embedded symbol, but we were visiting a referent every single time it was referenced!  That ends up traversing the grammar way too many times.

So we have to keep track of whether a tape has visited a particular symbol before.

Right now I'm remember what tapes have visited the EmbedGrammar in the EmbedGrammar itself, which required some changes to the algorithm.  In particular, we have to do the collection tape-by-tape now, rather than all at once.

Now that I'm writing this, it might have been better to remember the visits in the Tape itself, not in the EmbedGrammar.  Possible todo for the future.