Simplification of previous change to vocab
As mentioned in the last commit, it was a bit overkill to keep track of what symbols we had visited in the symbols themselves.  Also it's stateful, and since this is something we run several times, there's a possibility that state saved in the previous run somehow causes things to go wrong in the current run.  (I can't think of exactly how, but vocab is already hard enough to think about.)

I suggested keeping track of it in the tape instead, but this is also stateful.

Instead, we just keep this information (what symbols we've already visited) as a parameter to the call.  And since we never allowed a stack deeper than 1 recursion, this information also suffices to prevent over-recursion, so we can get rid of the stack argument.