Simpler output trees and bugfixes to vocab
Now that deriv is all concrete, the remaining use of BitSets was just to interface with the output trees, which still expected them.  This commit reimplements output trees to use strings, and meanwhile also goes back to the original, simpler solution where all tapes share the same tree.  (I had separated them out into tape-specific trees earlier, in anticipation of flag tapes, and then we never ended up using flags.)

This makes generation a bit quicker, either by avoiding the creation of bitsets or avoiding doing Object.assign() every time we add a token.

This change also revealed a bug, that DotExpr and DotStarExpr weren't checking whether the target was actually in the tape's vocabulary.  Targets that aren't in a tape's vocab can occur, like in certain matches (like match test #11).  This bug was probably hidden by going back to Tokens in the end, and having that quietly fix this up.

This commit also cleans up a bug in the previous commit, during vocab collection, where we were storing the tape name instead of the symbol name when remembering what symbols we'd already visited.