Bug in embed+hide
Embedding something that hides a tape was causing a bug, and we weren't testing for it!

In the end, it was one small error (assigning to "name" in HideGrammar's constructor where I should have had "this.name").  This had a result that every time we transformed the grammar, the hidden tape changed its name.

This commit also changes a few small interface things: for in-panel sampling, we only sample 1 thing now, rather than 5, and the hover/active color of buttons is darker so you know when it's thinking.