Bug in embed+hide, sheets QoL improvements
(1) Embedding something that hides a tape was causing a bug, and we weren't testing for it! In the end, it was one small error (assigning to "name" in HideGrammar's constructor where I should have had "this.name").  This had a result that every time we transformed the grammar, the hidden tape changed its name.

(2) Quality-of-life improvements for the sheets programmer:

  -- As things get slow for big grammars, sampling 5 results can be slow, and also once you get a lot of tapes the results are too hard to read anyway.  We now default to sampling 1 here.

  -- Buttons have darker shading when you hover/press them, so it's more noticeable when they're working.

  -- When sampling has zero results, we now show a message "No results found" instead of just an empty sample list.