Nested replace bugfix, tests for newfound bugs
Grammars with nested replacements weren't working; there were two problems, one of which was easily dealt with, and the other of which is still unclear:

(1) The combination of a hide and an embed (which happens in this configuration, but we weren't testing for in isolation) was bugged due to a missing "test." before one of the assignments in HideGrammar.constructor().  That's easily fixed.

(2) Beyond this, however, nested replacements on the same tape (e.g. where the programmer just says it's to/from "text" and we adjust the real names behind the scenes) aren't working.  I'm not yet sure why, especially as nested replacements ARE working with different-tape replacements, and we try to keep those structurally as similar as possible.  (For example, they still involve renames in the same configuration, it's just in different-tape replacements the renames are trivial.)

I've added several tests that show how the same- and different-tape nested renames fail and succeed respectively.  Warning, some such grammars will generate infinitely so put in a max character argument to testGrammar.