Dummy cells no longer singletons
Various things reference the positions of cells (e.g. Sheet1:4:12), including the names of the ad hoc tapes we use when doing a same-tape replace.  We just needed something unique to that replacement, and the cell position suffices.

However, when using the convenience functions for grammar construction (e.g. Seq, Replace, etc.), we just referred to a singleton cell DUMMY_CELL.  Since this always had the same position, all same-tape replacements end up using the same name for their tapes!

Now it's not a singleton; instead you use `new DummyCell()` and each one gets a unique address (based on a counter).

This doesn't solve the ultimate problem with nested replacements, but it's a necessary fix to write tests for them using the convenience functions (as opposed to writing them as sheet source).