Negated replace and trivial replace tests
(1) Found two new kinds of bugs, possibly related to each other and to the nested replace bug.

  (1a) Trivial replacements (like asking it to replace e->e) fail.

  (1b) Negating a replacement improperly excludes the case where the input is left alone.  (E.g., if we apply "t1:i->t2:a" to "hi", the result {t1:hi,t2:hi} is correctly excluded from the output.  However, this result SHOULD be generated by the negation of this grammar.)

This is suggesting to me that something isn't quite right with the way that we exclude nonreplacement cases.

(2) It was possible to generate infinitely (or effectively so) on hidden and other ad-hoc tapes (those beginning with "__"), because they weren't counted towards the character limit.  In this instance, the infinite generation I found might ultimately be due to a bug, but I can probably come up with a grammar that has this problem anyway, so we should count hidden characters towards the character limit.

However, as a result, the character limit appears to act a bit strangely from the POV of the client (because the total number of characters produced may be fewer the limit, due to some of them being hidden).  Maybe it should be a per-tape character limit instead.