Finding bugs due to non-minimal final exprs
Darlene had found at least one of the causes of our replace bugs: expressions that, although they correspond to outputs that should be yielded, haven't managed to simply to epsilon.  Doing it the other way around (not detecting success through minimalization to epsilon, but detecting through minimalization to null) seems to run into similar problems.

All of the non-minimalization examples I've seen so far involve negated epsilon, suggesting that that's ultimately our culprit.

I suspect that all final expressions that don't manage to minimize are bugs.  There are many such expressions, although many don't happen to make tests fail.  Probably they would if we switched over to searching for null rather than epsilon (whereas some tests that would succeed in the null search would have failed in the epsilon search).  That is to say, I suspect that for every non-minimized expressions like this, we could construct a test case that fails.

As a proof of concept of this idea, I found a non-minimizing formula in the VERBOSE logs, intersecting/joining/filtering epsilon with (epsilon | not-epsilon), and managed to create a failing test based on it.

Going back and reimplementing Expr.accepting() everything would probably handle all of these, but I can't beat the feeling that it's ultimately important to get minimization & negated epsilon right, so I'd rather keep exploring this avenue for now.

Other small fix: Embed(null) should simplify to null, but we weren't doing it.  This wouldn't make any current tests fail, but it would if we switched to null detection, and (like above) we probably could have found a failing test for it if we tried.