More graceful recovery from renaming conflicts
If the programmer erroneously asks for a rename where the destination tape name already exists in the child grammar, the resulting grammar would violate some otherwise solid invariances.

(Or even more importantly, violates an invariance that I'd like to be true, that every expression has reduced to either epsilon or null once every tape is delta'd out.  This isn't true when there's literal material that doesn't correspond to any tape known at the highest level -- we'll never do anything with that material because there will never be an deriv or delta on that tape.)

So now, we have a special transformation that goes through and fixes this up.  If it happens that a particular rename is invalid like, say:

    Rename(t1->t2, t1:hello+t2:world)

this is fixed up to:

    Rename(t1->t2, Rename(t2->__ERRt2, t1:hello+t2:world))

So the material in t2:world will eventually be consumed, although it won't be shown to the user in the end because it'll be on the tape __ERRt2. (Note that this makes it so that some erroneous grammars that had no outputs now have outputs.)