Non-sheet tests of startswith/endswith/contains
It turns out there were no simple tests of startswith/endswith/contains.

On that note, I determined that tests like endswith(hello, not(o)) are "failing" for the wrong reason.  We were constructing the relevant filter here:

   .* + ~(o)

But this is the wrong filter: the empty string is a member of ~(o), so the derivative of this (using a RTL derivative) includes dot-star.  (In other words, this doesn't filter anything.)

This was still getting no outputs, but because it eventually ends up with someting like (eps|~eps), which weren't properly reducing and therefore could not yield outputs at the end of generation.  eps[eps|~eps] should succeed, that's nullable!  When we properly reduce this, these tests start failing.