New approach to starts/ends/contains
In previous commits, I found that starts/ends/contains were sometimes not yielding outputs for a wrong reason (that they had encountered an unsimplifiable final expression at the end of generation).  This problem had been hidden for some time because all the test cases for this were semantically incorrect as stated, at least for the semantics of starts/ends/contains that we had been using.

The tests were like this:

   Input: StartsWith(hello, ~(h))
   Expected output: []

According to our semantics of StartsWith, the semantics of this should be hello[~(h).*].  But what I hadn't realized was that this really does have outputs, it outputs "hello".  Epsilon is a member of ~(h) and "hello" starts with epsilon.  "he" is a member of ~(h) and "hello" starts with "he", "hello" is a member of ~(h) and "hello" starts with "hello"!

We weren't outputting anything here because of a bug.  Now that this bug is addressed, we're outputting "hello" and a bunch of our tests break.

Our options are:

(A) The semantics is correct as we had been using it, and this really does have the output "hello".  The programmer using ~h in this way is incorrect about what they want, they don't want the complement of the language h, they want any letter but h.  So what they should have been doing is something like X&~h where X is a symbol representing any letter.  (Except that we don't yet support regexes that combine symbols and literal text, so they couldn't have actually said this.)

Downside: this is a breaking change so far as backwards compatibility is concerned, and we know that grammars already out there (Fineen's) use negation in the same way that our tests do.

(B) The tests really represent what we want, and we have to adjust the semantics to match it.  The semantics of that input should be hello[~(h.*)] instead.  That's probably what the programmer using "starts text:~h" expects!

This new branch attempts (B).  It introduces a new set of grammars (corresponding to the filter sides of starts/ends/contains), which add the .* where appropriate.  (That is, these grammars' constructExpr are basically doing what the old {StartsWith/EndsWith/Contains}Grammars did in their constructFilter functions.)  Then a transform changes their scope as appropriate, like:

   StartsWithFilter(Not(x)) -> Not(StartsWithFilter(x))

That gets the .* under the scope of the negation like the programmer intends.

This is implemented for a variety of regex operators that might occur in a filter regex, but I am skeptical that this approach will work in the general case (like for any regex).  I can already think of a few where it doesn't, in the current implementation:

  StartsWith(hello, ~(h)+ε)

In general, StartsWith(x+y) == x + StartsWith(y).  (That is, it adds its dots to, and thus after, the last child.  Don't want to add dots to/after the first child!)  But this results in the same problem we had originally, this will accept "hello".

Well, what if we're smart enough to ignore epsilon?  But then how do we handle:

  StartsWith(hello, ~(h)+(e|ε))

I can think of a solution here too... but it's very hard to do under our current architecture, and beyond that I am worried about how easy it is to come up with these counterexamples.

Then again, if someone is coming up with regexes this strange, they're pretty advanced already.  We could tell them, "If you want weird stuff like this, use equals and write the regex you really want.  Starts/ends/contains are convenience functions for beginners, where a somewhat simplistic algorithm constructs a starts/ends/contains regex for them.  We'll tell you the algorithm, and if it isn't what you want, roll your own filter."

Anyway, I'm going to continue down this path for a while, but in a branch in case we decide otherwise later.