Fixing tapes in scope-adjusted filters
The scope-adjustment algorithm described in the last commit wasn't calculating tapes correctly.  It turns out the XFilterGrammars need to know their tapes and remember them when being rearranged, otherwise our ordinary tape-inference algorithm can get it wrong (since their scope changed and this can change the tapes referenced within their scope).

Consider an example from last commit's algorithm:

  StartsWithFilter(x+y)
  -> x+StartsWithFilter(y)

This works if x and y reference the same tapes, but if what if they don't?

  StartsWithFilter(t1:h + t2:w)
  -> t1:h+StartsWithFilter(t2:w)
  -> t1:h + t2:w.*

But that's probably not what the programmer meant, they presumably wanted t1:h.* = t2.w.*.  (That is to say, they probably didn't write this meaning "t2 startswith w but t1 *equals* h"!)

So this StartsWithFilter object has to remember the tapes it started with (that is, when cloning the StartsWithFilter during the scope transformation, we have to assign it the same tapes the original had).  Since we do this so often, I added a tapes parameter to the constructor.