Fixing up some filter-epsilon problems
As mentioned a few commits ago, the presence of epsilons at the beginnings or ends of sequences can mess up the construction of filters that work like the programmer intends.  This isn't (in the general case) an easily addressable problem, but the more we can avoid this, the fewer surprises for the programmer.

One systematic source of epsilons (and thus a source of filter errors) is the table-parsing algorithm, specifically how we parse lines into sequences.  The recursive algorithm begins with the null case (epsilon), and so every sequence created in this way starts with epsilon.  This is fine in general (epsilon is concatenative identity so it doesn't matter), but it matters underneath an EndsWithFilter, since the special semantics of filters means that concatenation isn't really concatenation here.

We could just change the way table-parsing works, but I'd been meaning to "clean up" sequences after table-parsing anyway.  Some informal experiments on left/right branching sequences suggests it matters, depending on execution direction (LTR or RTL), but the table-parsing algorithm always creates left-branching sequences.  But this is a consequence of the algorithm, not something true of the grammar considered in the abstract.  (It also makes for hard-to-read string ids.)  In the end, of course, the ConcatExprs will be left/right branching because that's a binary operator, but this is an orthogonal issue.  Anyway, I've kind of wanted to flatten and clean up SequenceGrammars for a while now, and this was a good excuse to do so.

So now, SequenceGrammars are transformed during the transformation chain so that they're flat and don't contain EpsilonGrammars, and AlternationGrammars become flat and don't contain NullGrammars.