Removing StartsWithGrammar etc. as trivial
After the previous commits, StartsWithGrammar, EndsWithGrammar, and ContainsGrammar had become trivial.  All they had was an .id and an .accept() method for transformations, and the only thing that happened to them during transformations is to transform themselves into an EqualsGrammar with an XFilterGrammar as their child2.

So it's more straightforward to create them as that in the first place.  (This also gets rid of a slight annoyance.  Any Grammar created during a transformation doesn't have any tapes, but XFilterGrammars need to know their tapes, so the transformation of StartsWithGrammar etc. had to run calculateTapes on them after creating them.  This new way has them get their tapes in the ordinary way, at the same time as everyone else.)

Also, having fewer Grammar subclasses means less boilerplate code in the Visitor pattern.