Rectification of names
Three gradual changes led to some inconsistent naming conventions:

(A) The recent changes that eliminated StartsWithGrammar etc. and introduced StartsWithFilterGrammar etc.

(B1) Some time ago we changed the syntax of filters from "startswith text" to "starts text" etc., but a lot of things in the code were still named "StartsWith" etc.  It was getting a bit hard to remember which things had "with" and which didn't.

(B2) Some plain-filter things were called "Equals" (e.g. EqualsHeader) and others "Filter" (e.g. transformFilter).  (But meanwhile we had used filter in several different ways in the code, sometimes it meant the join-like operation and other times it meant the second child of that join, the one that specifies the condition that we're filtering for.)

(C) As we get more GrammarTransforms, some of them end in Transform but others don't.

So now:

- FilterGrammar -> EqualsGrammar (actually this was in the last commit)
- StartsWithFilterGrammar -> StartsGrammar
- EndsWithFilterGrammar -> EndsGrammar
- ContainsFilterGrammar -> ContainsGrammar

- .transformFilter() -> .transformEquals()
- .transformStartsWithFilter() -> .transformStarts()
- .transformEndsWithFilter() -> .transformEnds()
- .transformContainsFilter() -> .transformContains()

- EqualsHeader (stays the same)
- StartsWithHeader -> StartsHeader
- EndsWithHeader -> EndsHeader
- ContainsHeader (stays the same)

- Filter (convenience func) -> Equals
- StartsWith (convenience func) -> Starts
- EndsWith (convenience func) -> Ends
- Contains (convenience func) (stays the same)

- NameQualifier -> NameQualifierTransform
- ReplaceAdjuster -> SameTapeReplaceTransform
- RenameFixer -> RenameFixTransform
- FilterCreatorTransform -> FilterTransform