Refactored testMatchOutputs and added more testNotMatch tests.
Some tests in testNotMatch can have a large number of outputs, even for
a small test. In order for testMatchOutputs to handle this smoothly, we
split the output comparison into blocks, the size of which depends on
the number of expected/actual outputs. We also increased timeout from 2
seconds (2000ms) to 10 seconds (10000ms) in order to allow us to keep
the comparison blocks quite large.

We also added a generateOutputsFromGrammar function in testUtils to
assist in generating the expected outputs for tests in testNotMatch
because we often can write an alternate negation grammar for the
expected outputs.