Character counts handled by CountGrammar/Expr now
The maximum allowed characters of a generation (that is, the maximum depth of the tree created) used to be an argument to the generation function (through opt), and kept as part of the output/prev stack.

Now, there's a special Grammar and Expr that keeps track of this instead.  The benefit to this is that we no longer have to commit to this one strategy for character counting.  Darlene noted that it is hard to write certain tests when all you can say is that the *total* number of characters on all tapes is N; it would be better for these tests to say for each tape that it cannot exceed N.  It would be annoying to rewrite EVERY test this way, however.  Encapsulating character counting in a grammar/expr means that we can potentially have various types of character counting (or other kinds of safety grammars) that a test writer can wrap around a potentially dangerous grammar.

In this commit, I don't refactor any tests to work this way, though.  For the time being, the Interpreter just constructs the grammar to get the same results as the old way.  Including the little bug where you have to specify maxChars one higher than you really need :/ Once the tests are refactored to use Counts and opt.maxChars is out of the picture, we can remove this workaround from Interpreter and the little bug will go away.