Better interface to CountTapes
We agreed during today's meeting that we mostly want to use CountTapes with a single integer argument (e.g., 3 characters on each tape) rather than set the numbers for tapes individually.  So now the constructor for CountTapeGrammar reflects that, and there's a CountTape() convenience function for that, too.

This code is still untested.