Hidden tape stripping now in Interpreter
In the past, the responsibility for stripping hidden tapes (i.e. those that start with "__") has been located in various places -- in the client program, in the Output object, and most recently in the generate functions in Expr.

Now we want a parameter to strip or not strip hidden outputs.  (So that, for example, we can test them.)  The reasonable responsibility for this parameter is in Interpreter (e.g. as a parameter to generate()), and so it's easiest to also give Interpreter the responsibility for stripping them.

So Interpreter.generate/generateStream/sample/sampleStream all now take a stripHidden boolean parameter, defaulting to true.