Partial fix to major nested-replace bug
I found the bug that was causing most (but not all) of our same-tape nested replaces to fail.  It only required a small change to RenamedTape, re-wrapping the results of getTape() in a successor RenamedTape.  (Exactly as we do for matchTape(), actually.)  If the results of this aren't properly re-wrapped, it's possible to get a tape from getTape() that believes its local name is different from the name it's supposed to have.

For example, say you've got a MatchState looking to match t1 and t2.  However, t2 is really a rename of the tape that, globally, is called t3.  When you call MatchExpr.delta(t1) or MatchExpr.deriv(t1), it has to query its child on both t1 and t2.  (This is the only kind of Expr that ever has to query its child on a different tape than it itself was queried on.)  So to get a reference to t2, it calls t1.getTape(t2).  Now, t1 here references a RenamedTape. (Remember that a tape is a RenamedTape by virtue of passing through a RenameExpr, whether or not that renaming is relevant to the current tape.)  Since it's a rename tape, that request for .getTape(t2), when it's passed to the RenamedTape's child, gets changed to .getTape(t3).  That's correct, we need to be looking for t3 because that's what it's named globally, that's the only named that the global TapeCollection knows it as.

However, once we find that, if we just return that intact, we have a problem.  The problem is that when we ask this tape what its name is, the answer is "t3".  If this is returned like this to MatchExpr.delta(t1), we've got a problem, because in that scope, this tape should be named "t2".  So if we were to (say) pass this through another rename, say, t4->t2, things will go wrong because the rename will ask it "are you named t2?" and it'll say "nope".

So certain configurations of Matches and Renames can lead to trouble, and certain nested JoinReplaces happen to create those configurations.  Re-wrapping the returned tape (here, t3) properly so that it has its correct local name (here, t2) fixes that problem.

This is not the only problem; there is still a nested sane-tape JoinReplace that should work, but doesn't.