Refactoring namespaces
One medium-term goal we have is to allow the declaration of namespaces by the programmer, like by a command "namespace:".  (Real human programmers typically aren't going to be using this, but I want this command to help us auto-generate single-document projects.)

But our current namespaces would make this tricky, because of a tiny implementation detail:
they know their own name.  When we're only creating them in code, this is fine (we always
know this name at the point we have to create them), but in the TST-parsing algorithm, this is information that's hard to access at the point the namespace object would be constructed, and might not even exist.  (Like if the programmer doesn't actually assign the namespace to anything.  This is probably an error that we want to flag, but we still need to construct the grammar.)

This commit refactors the namespace object and name-qualification transformation so that namespaces do NOT know their own names.  The only part of this that's non-trivial is the way the name qualification algorithm creates a stack of namespaces that provides the context that embeds use to resolve what symbol they refer to.  This stack needs both references to the namespace objects and their names.  When namespaces knew their names, this was easy: we only had to maintain a stack of the namespaces themselves, and when the transform visited the namespace, it added itself to this stack.  Now, a namespace doesn't know its name; only their parent namespace knows its name.  So now the stack contains [name, ns] pairs instead, and it's the responsibility of the parent namespace to add them to the stack instead.

Two other little changes:

(1) Some of the transforms assume that they start with a NsGrammar, and if they don't, it's their responsibility to wrap whatever they get in a trivial NsGrammar so this is true.  This commit moves this responsibility to Interpreter itself.  Interpreter.constructor checks the passed-in grammar and wraps it in an NsGrammar if necessary before executing any transforms.  Now transforms can assume the root is an NsGrammar, and the type of .transform now reflects this.

(2) Commentary on each transform detailing what it does and why.