Namespace: operation
This commit makes the "namespace:" operation available to programmers.  Now, groups of assignments can be surrounded in a namespace enclosure, which can (and should) in turn be assigned to a symbol.  This then changes the fully-qualified name with which you can refer to the symbols in that namespace.

For example:

X:, namespace:, root:, text
  ,           ,      , foo

MAIN:, embed
     , X.root

Note that just "root" will no longer work here; there is no longer a symbol called "root" in the sheet-level namespace to which MAIN belongs.

I don't expect that human programmers will get much use of this one, it's more for automatically generating single-sheet source files for the convenience of client interface developers.