Single-sheet source packaging and download
The last few commits created a syntax for the creation of namespaces within source files.  In this commit, we use that ability to "package" a multi-sheet project into a single source file, and create a button in the GSuite interface to download the result.

This will make things easier for developers of client interfaces.  Now, rather than have to understand/manage Gramble's multi-source loading loop (or roll your own), you can load/embed a single file with the entire project in it.