ReplaceGrammar can take prespecified tape names
When testing the results of Replaces (where a same-tape conflict involves quietly renaming one of the tapes), it would be difficult to test the contents of those hidden tapes because the name of those tapes is difficult to determine in advance.

ReplaceGrammar.constructor and Replace now take an argument hiddenTapeName, which lets the test programmer specify that they want that tape renamed to something specific.  (NB: This only ends up being used if the replace is a same-tape replace.  Different-tape replaces don't involve renaming tapes.)