Added the include-hidden-tape negation tests missing from testNotReplace (2c,5c).
Also fixed transformReplace to pass the hiddenTapeName to the new ReplaceGrammar,
and ensured that the hiddenTapeName in ReplaceGrammar starts with a double
underscore ("__").

The tests implemented are those referred to in commit eab4d9cf62bd2dd57bd7bd8e039f5bb835b6a360:

Commenting out not-yet-true Negation tests
I'm coming to the conclusion that we may need two different kinds of negation:

(1) Negation relativized to a set of tapes that includes hidden tapes.
This is what we already have.

(2) Negation relativized to a set only containing visible tapes.  We're
going to need this for some future constructions.

Our failing notReplace tests were failing because we had implemented (1)
but were testing for (2).  I had suggested that we fix the results to be
consistent with the true results of (1), but now I'm thinking this
should be a newly-written test instead, and we should keep around the
existing test for when (2) is implemented.

Anyway, I commented them out for the time being, because it's not clear
when we'll get around to implementing the appropriate kind of Negation.