Operator parsing framework
Operators and assignments (e.g. "table:", "VERB:") were one of the last kinds of text where we were doing ad-hoc parsing (i.e. parsing in code rather than with a MiniParser grammar).  This was fine because all operators/assignments were atomic.  However, the changes to rule syntax discussed in the previous two meetings mean that operators are going to become more complicated.

The new syntax for same-tape replacement rules doesn't use headers to name the affected tape anymore (because that effectively prevents rule embedding). Now the affected tape has to be specified somewhere else... and the operator is the only place to put it.

So we went from this:

   replace:, from text, to text ...

to this:

   replace text:, from, to

Rather than do this in an ad-hoc manner, it's finally time to parse operators in a more principled manner.  Now, operators/assignments are parsed using the same machinery as headers and regexes, using MiniParser.