Implementing replacement cascades
There is now a new kind of replacement rule (that will presumably be the main kind that most programmers will use), in which the input/output tapes have fixed names (just "input" and "output"), and when joining to the grammar we can rely on those names.  This restriction makes some things possible that were difficult or impossible before, like assigning a rule to a symbol and embedding it, and having a block of rules apply one after another.

Now the syntax is like this:

replace text:, from, to
             , x   , y

The signal for the compiler to use this new kind of rule is the operator "replace X:", which creates a different kind of Op, which creates a different kind of TST, all the way down the line.

The names of classes are up in the air at this point, but to distinguish the two kinds, for the time being I've just been using the name "rule" instead of "replace".  I'll sort out consistent naming later.

The only part of the compilation that's non-trivial here lives in the file ruleReplace.ts, which takes a JoinRuleGrammar and turns it into the appropriate tree of renames, joins, and hides.  The logic is ultimately a lot simpler than what's going on in sameTapeReplace.ts.