Changed prefix for hidden tape names
Previously, starting a tape name with "__" hid it from interfaces.  However, since it's entirely possible for the programmer to name something starting with "__", that can have unpredictable errors and hard-to-debug side effects in the rare case they happen to choose the same name as a tape that's hidden.  Programmers probably shouldn't be allowed to hide tapes in this way in the first place -- they should use "hide".

So either we forbid (in the tabular syntax) tape names from starting with "__", or we choose something to be impossible.  Most programming languages allow names starting with "__", and it's also a permissible character inside the word, so we chose the latter.  Now the hidden tape prefix is ".".

Also, the hidden tape prefix is now a global constant, as it always should have been but wasn't.