Fixed bug with end-of-line empty cells
A few more tests for in-sheet unit tests revealed a bug, that sometimes an empty string in a content cell wasn't seeming to be taken into account in the test.  (E.g., the test should have asserted that there was an empty string on a particular tape, but it wasn't asserting anything of that tape.)

The culprit ended up being in the way we handle CSVs.  Our CSV parsing results in lines of variable length:

[
[ "%% comment " ],
[ "" ],
[ "VERB:", "table:", "text", "gloss" ],
[ "", "", "omo", "jump" ],
[ "", "test:", "text", "gloss"],
[ "", "", "omo", "jump" ],
[ "", "", "omo" ],
]

Just looking at this file in a spreadsheet editor, you might think the second test in this program (on the last line) should fail. There's an empty cell on the last line under "gloss", so the test should be looking for the empty string in the gloss tape, and there's no output consistent with that, so the test should fail.  But the test succeeds, because underneath we're just iterating over these strings, and there's no string there, so nothing is ever made.

Now, instead, when we encounter a line that is shorter than the maximum line length so far encountered, we treat it as if it's padded out with "".  So now the last line is [ "", "", "omo", "" ] and the test fails as intended.