Integration of regexes into new rule syntax
A bunch of updates related to the new rule syntax:

(1) We hadn't been testing pre/post parameters in the new rules yet.  Now we are.

(2) We hadn't yet allowed regexes in the new rule syntax (e.g., "from re" with an implicit tape name).  With the old way we were parsing implicit tape names, this would be a bit difficult because the implicit tape name changes depending on from/to, but we don't know that at the point we create the "re".

This commit eliminates the special atomic from/to/pre/post classes (e.g. AtomicFromHeader) in favor of constructing these directly in the parser functions.  (So rather than create an AtomicFromHeader(), we create an equivalent header TagHeader("from", TapeNameHeader(".input")).)  This way lets the parser also look for longer phrases of interest, like "from re", and construct the appropriate header for that.

(3) We hadn't been testing the combination of from/to/etc. and regexes in the tabular syntax.  (It worked, fwhew, but we hadn't been testing it.)  This adds such tests in both the old-style and new-style rule syntax tests.

(4) As agreed in the meeting ~2 weeks ago, the input/output tapes of the new-style rules should be hidden (i.e., be ".input" and ".output" rather than plain "input" and "output").  Otherwise, the programmer would get hard-to-understand bugs if they were to name a tape "input" or "output".  A few commits ago, I made globals for these names, but had never actually changed the code to use them.  Now it's changed.

(5) We hadn't been recognizing "re" as a reserved word, leading to the possibility of someone naming a tape "re".  This made "from re" ambiguous.